Don't drain command queue prematurely
If the client is configured to never send commands to slaves but it connects to a slave first, the
queue will be prematurely drained, causing "couldn't get client" errors to be sent to all queued
callbacks because getRandomConnection() will return null.

This commit changes the behavior so that the queue is only drained on `fullReady`. I believe this is
desired behavior, anyways, as opposed to draining the queue when one master is available because
otherwise we're ignoring the maxQueueLength setting and pushing the queuing on to the `redis` client
for all connections to masters that aren't ready yet.

This has the side effect of not blocking getSlots() from returning while the queue is drained but
I do not believe this is an issue in practice.